refactor(api/workflow): Move SystemVariable to `workflow/enums`.
