+++ api/core/app/segments/__init__.py
@@ -2,7 +2,6 @@ from .segment_group import SegmentGroup
 from .segments import (
     ArrayAnySegment,
     ArraySegment,
-    FileSegment,
     FloatSegment,
     IntegerSegment,
     NoneSegment,
@@ -13,11 +12,9 @@ from .segments import (
 from .types import SegmentType
 from .variables import (
     ArrayAnyVariable,
-    ArrayFileVariable,
     ArrayNumberVariable,
     ArrayObjectVariable,
     ArrayStringVariable,
-    FileVariable,
     FloatVariable,
     IntegerVariable,
     NoneVariable,
@@ -32,7 +29,6 @@ __all__ = [
     'FloatVariable',
     'ObjectVariable',
     'SecretVariable',
-    'FileVariable',
     'StringVariable',
     'ArrayAnyVariable',
     'Variable',
@@ -45,11 +41,9 @@ __all__ = [
     'FloatSegment',
     'ObjectSegment',
     'ArrayAnySegment',
-    'FileSegment',
     'StringSegment',
     'ArrayStringVariable',
     'ArrayNumberVariable',
     'ArrayObjectVariable',
-    'ArrayFileVariable',
     'ArraySegment',
 ]
