+++ api/core/app/segments/types.py
@@ -10,8 +10,6 @@ class SegmentType(str, Enum):
     ARRAY_STRING = 'array[string]'
     ARRAY_NUMBER = 'array[number]'
     ARRAY_OBJECT = 'array[object]'
-    ARRAY_FILE = 'array[file]'
     OBJECT = 'object'
-    FILE = 'file'
 
     GROUP = 'group'
