+++ api/core/workflow/entities/node_entities.py
@@ -4,13 +4,14 @@ from typing import Any, Optional
 
 from pydantic import BaseModel
 
-from models.workflow import WorkflowNodeExecutionStatus
+from models import WorkflowNodeExecutionStatus
 
 
 class NodeType(Enum):
     """
     Node Types.
     """
+
     START = 'start'
     END = 'end'
     ANSWER = 'answer'
@@ -44,34 +45,11 @@ class NodeType(Enum):
         raise ValueError(f'invalid node type value {value}')
 
 
-class SystemVariable(Enum):
-    """
-    System Variables.
-    """
-    QUERY = 'query'
-    FILES = 'files'
-    CONVERSATION_ID = 'conversation_id'
-    USER_ID = 'user_id'
-    DIALOGUE_COUNT = 'dialogue_count'
-
-    @classmethod
-    def value_of(cls, value: str) -> 'SystemVariable':
-        """
-        Get value of given system variable.
-
-        :param value: system variable value
-        :return: system variable
-        """
-        for system_variable in cls:
-            if system_variable.value == value:
-                return system_variable
-        raise ValueError(f'invalid system variable value {value}')
-
-
 class NodeRunMetadataKey(Enum):
     """
     Node Run Metadata Key.
     """
+
     TOTAL_TOKENS = 'total_tokens'
     TOTAL_PRICE = 'total_price'
     CURRENCY = 'currency'
@@ -84,6 +62,7 @@ class NodeRunResult(BaseModel):
     """
     Node Run Result.
     """
+
     status: WorkflowNodeExecutionStatus = WorkflowNodeExecutionStatus.RUNNING
 
     inputs: Optional[Mapping[str, Any]] = None  # node inputs
